reposource: Rename go_packages.go to go_modules.go
