+++ internal/conf/reposource/go_modules.go
@@ -8,18 +8,20 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 )
 
-// GoDependency is a go module at a specific version.
+// GoDependency is a "versioned package" for use by go commands, such as `go
+// get`.
+//
+// See also: [NOTE: Dependency-terminology]
 type GoDependency struct {
 	mod module.Version
 }
 
-// ParseGoDependency parses a string in a 'module@version' format into a GoDependency.
-func ParseGoDependency(dependency string) (*GoDependency, error) {
+// ParseGoModDependency parses a string in a '<name>@<version>' format into an
+// GoDependency.
+func ParseGoModDependency(dependency string) (*GoDependency, error) {
 	var mod module.Version
-
 	if i := strings.LastIndex(dependency, "@"); i == -1 {
 		mod.Path = dependency
-		mod.Version = "latest"
 	} else {
 		mod.Path = dependency[:i]
 		mod.Version = dependency[i+1:]
@@ -29,37 +31,37 @@ func ParseGoDependency(dependency string) (*GoDependency, error) {
 	if err != nil {
 		return nil, err
 	}
-
 	return &GoDependency{mod: mod}, nil
+}
+
+func (d *GoDependency) Scheme() string {
+	return "go"
+}
+
+// PackageSyntax returns the name of the Go module.
+func (d *GoDependency) PackageSyntax() string {
+	return d.mod.Path
+}
+
+// PackageManagerSyntax returns the dependency in Go syntax. The returned string
+// can (for example) be passed to `go get`.
+func (d *GoDependency) PackageManagerSyntax() string {
+	return d.mod.String()
+}
+
+func (d *GoDependency) PackageVersion() string {
+	return d.mod.Version
 }
 
 // RepoName provides a name that is "globally unique" for a Sourcegraph instance.
 //
 // The returned value is used for repo:... in queries.
-func (m *GoDependency) RepoName() api.RepoName {
-	return api.RepoName("go/" + m.mod.Path)
+func (d *GoDependency) RepoName() api.RepoName {
+	return api.RepoName("go/" + d.mod.Path)
 }
 
-// PackageSyntax returns the module name.
-func (m *GoDependency) PackageSyntax() string {
-	return m.mod.Path
-}
-
-// PackageManagerSyntax returns the module name and version in mod@version format.
-func (m *GoDependency) PackageManagerSyntax() string {
-	return m.mod.String()
-}
-
-func (m *GoDependency) Scheme() string {
-	return "go"
-}
-
-func (m *GoDependency) PackageVersion() string {
-	return m.mod.Version
-}
-
-func (m *GoDependency) GitTagFromVersion() string {
-	return "v" + m.mod.Version
+func (d *GoDependency) GitTagFromVersion() string {
+	return "v" + d.mod.Version
 }
 
 func (d *GoDependency) Equal(other *GoDependency) bool {
